==> notebooks/0_Crear tileset.ipynb <==
# Creación de tilesets
En este notebook vamos a crear como crear tilesets. Un tileset es un formato de archivo aceptado por mapbox que incluye información sobre los datos, la geometría de cada uno de los elementos, y otras opciones como el nivel de zoom permitido para visualizar.

import os

current_dir = os.path.dirname(os.getcwd())

año = 2019

Para crear el json de abajo hay que ir primero a https://www.ine.es/ss/Satellite?L=es_ES&c=Page&cid=1259952026632&p=1259952026632&pagename=ProductosYServicios%2FPYSLayout, seleccionar un año (2020) en Ficheros nacionales de arriba, y darle a "Ir". Esto bajará un zip, que lo cargamos en mapshaper. Esto lo hacemos automáticamente.

!mkdir {current_dir}/datos/mapas_elecciones

!cd {current_dir}/datos/mapas_elecciones && wget https://www.ine.es/prodyser/cartografia/seccionado_{año}.zip && unzip seccionado_{año}.zip -d seccionado_{año}

path_to_json = current_dir + f'/datos/mapas_elecciones/seccionado_{año}.json'
path_to_shp = path_to_json.replace('.json', f'/SECC_CE_{año}0101.shp')
path_to_mbtile = path_to_json.replace('.json', '.mbtiles')

El archivo json tiene que haber sido pasado por mapshaper, para corregir fallos en la exportación del json.
Para ello, en la consola escribimos `-clean`, y `-proj wgs84` para pasar a proyección Mercator.

# Las acciones deben ir en ese orden (-o el último) porque si no lo exporta y no aplica el resto de acciones!
!mapshaper -i {path_to_shp} -proj wgs84 -clean -o {path_to_json}

Una vez hecho esto, vamos a crear el mbtile para poderlo cargar en Mapbox studio. Para ello vamos a establecer un nivel de zoom adecuado. De z3 a z10 es apropiado.

!tippecanoe -o $path_to_mbtile --maximum-zoom=10 --minimum-zoom=1 --read-parallel --force $path_to_json

Este mbtiles se sube directamente a mapbox, y es el que se usará para hacer los análisis.
